requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # Para recibir archivos

# Generación de PDFs
weasyprint==60.2
Pillow==10.2.0  # Procesamiento de imágenes
pydyf==0.8.0

# Templates
Jinja2==3.1.3

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Email (opcional, para enviar PDFs)
# aiosmtplib==3.0.1
